requirements.txt
# Project dependencies
numpy
pandas
scikit-learn
streamlit
altair
requests
python-dotenv
google-generativeai
